Use entrypoint-service rockcraft property
The metallb charms do not explicitly set the entrypoint, expecting
it to be the default one.

For this reason, pebble receives the metallb arguments and errors out.

To avoid this issue, we'll need to define the entrypoint-service
rockcraft property.

Note that we're adding the [ -h ] arg to our service command as
rockcraft forces us to define an optional command arg.