Changes rock bases to bare (#6)
Switching to a bare-based image will reduce the overall image size
and reduces attack surface area.

The original Dockerfile uses gcr.io/distroless/static:latest, which has
only a few packages. We can add those packages and switch to a
bare-based image.

Helm charts may expect the speaker image to be able to copy some files.

The frr rock base is not switched to bare as it has quite a few runtime
dependencies and it is a bit more complex.

We can no longer use ensure_image_contains_paths to check if files exist
in the rock images, since they are now bare-based. Instead, we can use
ensure_image_contains_paths_bare, which checks the image layers instead.
Because of this, we need sufficient permissions to check the
/var/lib/docker folder.