fix: pin FIPS Go toolchain in override-build step (#18)
Rockcraft overrides the Go version during the Pebble build
when `GOTOOLCHAIN` is set to `local`,
which causes the binaries to be built with a non-FIPS
toolchain and breaks compliance.

In addition, the Go FIPS tracks `1.22` and `1.23` appear to
contain a bug that prevents proper FIPS enforcement.
To ensure compliance, the build now explicitly uses the `1.24` FIPS Go version.